requirements.txt
altgraph==0.17.4
numpy==2.0.0
opencv-python-headless==4.10.0.84
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pip-review==1.3.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pywin32-ctypes==0.2.2
pyzbar==0.1.9
six==1.16.0
watchdog==4.0.1
